fixed tests to pass and small refactor of password methods.
The tests needed a fix because of the change in the user model. Since all of the user models were being generated from a single function, so the refactor was extremely easy.

The other change was a code cleanliness refactor to go away from handling raw bytes, as well as changin the incoming data from  to .