Huge extension/refactor of `UserIdentifier` and `UserRegistrationForm`.
Involved changing the validator and fields in the `UserIdentifier` class
as well as refactoring the util functions that deal with users and user
identifiers.

Also changed the inheritance of some fields so the `User` model really
only serves as the interface for direct database operations.

Refactored/fixed some test code (easily due to fixtures!) so they could
pass with the new changes.